fantasy-sports-clustering-algos
The purpose of the project is to analyze past season data from the English Premier League to conduct a cluster analysis of football players' performances and potential. This analysis aims to use clustering algorithms to identify patterns in player performances and fantasy returns, enabling the OnSports fantasy sports platform to determine the appropriate starting prices for players in the new season. This helps the platform optimize player pricing based on their potential and past performance, enhancing the competitiveness and strategic elements of the fantasy football experience.