Business Case: Using Data Science to Optimize CX
The project aims to bolster AllLife Bank’s competitive stance in the credit card market by leveraging data-driven insights to improve market penetration and customer satisfaction. Recognizing the need for a strategic approach, the bank plans to implement personalized campaigns to attract new customers and enhance upselling opportunities among existing ones. Additionally, to address the prevailing negative customer perception of support services, the project focuses on overhauling the service delivery model to expedite query resolution and elevate the overall customer experience.

This initiative involves the "Data Science team" collaborating with the heads of marketing and operations to optimize campaign targeting, refine customer engagement strategies, and streamline support services through the use of a sophisticated machine learning model. The integration of these efforts underscores a commitment to innovation and customer-centricity, setting the stage for a transformative and successful financial year ahead.